binary search for java
