+++ C++Codes/gcd.cpp
